Click with positions works now.

+++ src/main/java/com/germaniumhq/germanium/points/Point.java
@@ -1,27 +1,27 @@
 package com.germaniumhq.germanium.points;
 
 public class Point {
-    private float x;
-    private float y;
+    private int x;
+    private int y;
 
-    public Point(float x, float y) {
+    public Point(int x, int y) {
         this.x = x;
         this.y = y;
     }
 
-    public Point adjustX(float x) {
+    public Point adjustX(int x) {
         return new Point(this.x + x, this.y);
     }
 
-    public Point adjustY(float y) {
+    public Point adjustY(int y) {
         return new Point(this.x, this.y + y);
     }
 
-    public float getX() {
+    public int getX() {
         return x;
     }
 
-    public float getY() {
+    public int getY() {
         return y;
     }
 }
